Divided code into separate files (in my opinion improves readability)
